+++ src/pages/Root.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { useRecoilValue } from 'recoil';
 import { Outlet } from 'react-router-dom';
-import authAtom from '../_state/auth';
+import { authAtom } from '../_state/auth';
 import Nav from '../components/Nav/Nav';
 import Login from './Login/Login';
 import { PageWrap } from '../styles/GlobalStyles';
@@ -13,7 +13,6 @@ const PageWrapStyle = styled.div`
 
 export default function Root() {
   const auth = useRecoilValue(authAtom);
-  // console.log(auth);
 
   return (
     <div>
